+++ clang-tidy/nodecpp/NodeCppTidyModule.cpp
@@ -13,13 +13,16 @@
 #include "ArrayTypeCheck.h"
 #include "MayExtendDeclCheck.h"
 #include "MayExtendLambdaCheck.h"
-#include "NakedPtrAssignmentCheck.h"
+#include "NakedAssignmentCheck.h"
 #include "NakedPtrFuncCheck.h"
 #include "NakedPtrReturnCheck.h"
+#include "NakedStructCheck.h"
 #include "NewExprCheck.h"
 #include "NoCastCheck.h"
 #include "PtrArithmeticCheck.h"
+#include "RawPointerAssignmentCheck.h"
 #include "StaticStorageCheck.h"
+#include "StdFunctionCheck.h"
 #include "VarDeclCheck.h"
 
 namespace clang {
@@ -36,20 +39,26 @@ public:
         "nodecpp-may-extend-decl");
     CheckFactories.registerCheck<MayExtendLambdaCheck>(
         "nodecpp-may-extend-lambda");
-    CheckFactories.registerCheck<NakedPtrAssignmentCheck>(
-        "nodecpp-naked-ptr-assignment");
+    CheckFactories.registerCheck<NakedAssignmentCheck>(
+        "nodecpp-naked-assignment");
     CheckFactories.registerCheck<NakedPtrFuncCheck>(
         "nodecpp-naked-ptr-func");
     CheckFactories.registerCheck<NakedPtrReturnCheck>(
         "nodecpp-naked-ptr-return");
+    CheckFactories.registerCheck<NakedStructCheck>(
+        "nodecpp-naked-struct");
     CheckFactories.registerCheck<NewExprCheck>(
         "nodecpp-new-expr");
     CheckFactories.registerCheck<NoCastCheck>(
         "nodecpp-no-cast");
     CheckFactories.registerCheck<PtrArithmeticCheck>(
         "nodecpp-ptr-arithmetic");
+    CheckFactories.registerCheck<RawPointerAssignmentCheck>(
+        "nodecpp-raw-pointer-assignment");
     CheckFactories.registerCheck<StaticStorageCheck>(
         "nodecpp-static-storage");
+    CheckFactories.registerCheck<StdFunctionCheck>(
+        "nodecpp-std-function");
     CheckFactories.registerCheck<VarDeclCheck>(
         "nodecpp-var-decl");
   }