+++ test/nodecpp-checker/Inputs/nodecpp.h
@@ -1,0 +1,24 @@
+#ifndef NODECPP_H
+#define NODECPP_H
+
+namespace nodecpp {
+template <class T>
+class naked_ptr {
+	T* ptr;
+public:
+	naked_ptr(T* ptr = nullptr) :ptr(ptr) {}
+
+	naked_ptr(const naked_ptr&) = default;
+	naked_ptr(naked_ptr&&) = default;
+	naked_ptr& operator=(const naked_ptr&) = default;
+	naked_ptr& operator=(naked_ptr&&) = default;
+
+ 	T* operator->() { return ptr; }
+	T& operator*() { return *ptr; }
+};
+}
+
+
+
+
+#endif NODECPP_H